tools/requirements.txt
antlr4-python3-runtime==4.8
attrs==20.3.0
certifi==2020.12.5
chardet==4.0.0
idna==2.10
jsonschema==3.2.0
mdutils==1.3.0
pyaml==20.4.0
pyrsistent==0.17.3
pytz==2020.5
PyYAML==5.4.1
requests==2.25.1
simplejson==3.17.2
six==1.15.0
SQLAlchemy==1.3.22
stix2==2.1.0
stix2-patterns==1.3.2
taxii2-client==2.2.2
urllib3==1.26.2
